New post: Flashing Motorola Droid to Cricket
Additionally, fixed small bug where `text` was different in paragraphs
and in lists. Added a line to main.css so `text` in lists looked the
same as if they were in <p />.

Now that I fixed that, I went back to the RNS-510 post and replaced a
few **text** items with `text` so it looks better. Also made the
references at the bottom clickable.