New feature: Last updated and series.
Added two new features.

1) Each post has two date-stamps. The one in the filename (page.date)
and the one in the front-matter (page.modified). It looks like page.date
is used for everything, which presumably is the original posted date,
while page.modified is only used in RSS. The new _layouts/post.html file
will display "This page was last updated on June 29, 2013" (example) at
the top of any post who's page.modified and page.date differ. Users can
turn this off by adding "hideupdated: true" in the front-matter.

2) Post categories also seemed little-used. This modified file will now
display all the pages with the same category set at the top of the post.
By default this is disabled so the user will need to set "showseries:
true" in their front-matter to enable it. The series listing will show
links to other pages of the same category, and non-linked text for the
current page. This is only tested with single-category pages at this
time.