Improved "last updated" and "series", re-dated too.
I simplified "last updated" based on Michael Rose's suggestion. Instead
of normalizing dates and comparing differences, last updated will show
if page.modified is set. I didn't think that it was not required in the
front-matter to begin with. Now, if the user specifies it, the post will
show the last updated date above the content, even if it's the same
date.

Series was just a small change. If it's the same page being viewed,
instead of just the non-linked title, it will be a non-linked title in
<strong />. Without <strong /> it would probably confuse the user why
they cannot click on it before they realized it's the same page they're
looking at.

The remaidner of the changes are re-dated filenames and modified dates.
I went through all of these pages and corrected the original and
modified dates, and since I am renaming the files, I went ahead and used
better URLs for the posts. They are now the same as the post title.
Looks a lot better. No content change at all.