Image annotations and improved series.
Turns out adding the border CSS to <img /> breaks alignment. I had to do
that in the parent div (annotparent). Also implemented image annotations
(text overlaid ontop of images).

I've also improved the Series feature. When there is only one post in
the same category, the series div wont be displayed even if the user
requests it. Only if there are a minimum of 2 posts will it actually
appear if the user wants it.