Moving tracking out of CloudFlare, manual redirect
Google Analytics app in CloudFlare isn't really that good. It wasn't
injecting in the redirect html files, so I was missing most of my
referrals. Moving it back into the repo, and including it into the
layout in hopes of it surviving minification.

Not using redirect plugin, I have no control over the redirect template.
Doing redirects myself.