Handling edge-case: COMPATIBILITY=false
When updating the kernel with COMPATIBILITY=false in resize2fs,
initramfs-rebuild will build for both kernels and the last one will
remain as the only one in initramfs.gz regardless of the actual
hardware. When updating the kernel the post-install hook is called for
each kernel.

Handing this by having initramfs-rebuild skip building on the wrong
hardware. If COMPATIBILITY=true then both kernel's modules will be in
the final image, and when false only the current hardware's modules will
be present.

Tweaking shell scripts for consistency and less lines as per examples in
https://wiki.ubuntu.com/DashAsBinSh and
http://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html

Putting some content in 404.rst instead of the dumb message I've had for
so long.