Create pre-releases on PR merges. (#91)
* Create pre-releases on PR merges.

To be used by new diff.yml GitHub Action on the way. Using pre-releases
to store HTML files that can be diffed in PRs. Currently diff is against
prod, which gets ugly when merges to main branch aren't immediately
published.

Pre-release titles will be the same as the merged PR's title and bodies
are the current "Unreleased" section in the changelog file.

* Rename tag.

So it doesn't get annoying with my zsh/git tab completion.