Adding placeholder blog post.
I don't have content for a blog post yet but I don't want to forget how
to make posts using Sphinx. So for now I'm adding a hidden blog post
that serves as a placeholder. It still shows up on the sitemap but not
on the front page or any sidebar.

To make a blog post uncomment the stanza in .toc.yml and remove the
field list items at the top of the md file.