Disable Atom feeds (#272)
Enabled this just because it was available in #267 but turning it off
now. XML files have an updated field with the current time of
generation. Not based on mtime of any file, running `make clean docs`
confirms the current date is set. This causes my diff feature in PRs to
be nearly useless.