Making docs version static for Sphinx
Main driver for this change is to reduce noise in HTML diffs in pull
requests. Keeping the version static in the documentation produces
deterministic output files.

This wasn't a problem before I upgraded Sphinx and other libraries to
the latest version, which now inject the docs version in
documentation_options.js This changes the hash which is now hard-coded
into HTML files. So every PR causes changes to ALL HTML files.