Rewording, reducing the size of the influxdb v2 diff (#305)
* Add png extension.

So the linked URL goes to the png file instead of the default jpg. All
of these are PNG screenshots.

* explain what TIG does

* Deduplicate, no summary.

Might be confusing and the user might think these two bullet points are
actions they have to take right then and there.

* linux_amd64, 10s, move into note

Using "linux_amd64" so the user can just copy/paste into the browser
search field to find the right file.

Setting telegraf to the normal 10s interval. I've found a way to
implement downsampling so we can collect the normal amount of metrics
now.

Removing note about matching the interval to the config file. Link to
the config file may confuse the user.

Moving the alert explanation into the note admonition. Not relevant for
setting up the alerts, only for the curious.

* Monitoring docker

Collecting docker metrics, since TrueNAS SCALE uses it to run apps.

* Rephraising grafana instructions.

I think it's more clear now.